SI-4492 feat(notes): Отдельный слой провайдера
