+++ draggable_notes/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:draggable_notes/di/di_container.dart';
 import 'package:draggable_notes/interactor/notes/hive/hive_storage.dart';
+import 'package:draggable_notes/providers/theme_provider.dart';
 import 'package:draggable_notes/res/themes.dart';
 import 'package:draggable_notes/storage/notes/hive/adapters/note.dart';
 import 'package:draggable_notes/storage/theme/prefs_theme_storage.dart';
