+++ draggable_notes/lib/res/themes.dart
@@ -1,5 +1,4 @@
 import 'package:draggable_notes/res/colors.dart';
-import 'package:draggable_notes/storage/theme/theme_storage.dart';
 import 'package:flutter/material.dart';
 
 class AppThemes {
@@ -44,27 +43,6 @@ class AppThemes {
       bodyMedium: TextStyle(color: AppColorsDark.textColor),
     ),
   );
-}
-
-/// Провайдер для темы
-class ThemeProvider extends ChangeNotifier {
-  /// Хранилище темы.
-  final ThemeStorage _themeStorage;
-
-  /// Текушая тема.
-  late ThemeMode _currentTheme;
-
-  ThemeMode get currentThemeMode => _currentTheme;
-
-  ThemeProvider(this._themeStorage) {
-    _currentTheme = _themeStorage.getThemeMode();
-  }
-
-  set currentThemeMode(ThemeMode value) {
-    _currentTheme = value;
-    _themeStorage.setThemeMode(_currentTheme);
-    notifyListeners();
-  }
 }
 
 extension StringTheme on String {
